add-contextual-data: get rid off preliminary separation of CSV/non-CSV scanners
The code as is today is prepared for adding new formats to read data
from. As new data formats from files is not anticipated, this patch
gets rid of that complexity.
