+++ modules/add-contextual-data/CMakeLists.txt
@@ -10,8 +10,6 @@ set(add_contextual_data_SOURCES
     contextual-data-record.c
     contextual-data-record-scanner.h
     contextual-data-record-scanner.c
-    csv-contextual-data-record-scanner.h
-    csv-contextual-data-record-scanner.c
     add-contextual-data-selector.h
     add-contextual-data-template-selector.h
     add-contextual-data-template-selector.c
