+++ modules/add-contextual-data/contextual-data-record-scanner.c
@@ -21,53 +21,136 @@
  */
 
 #include "contextual-data-record-scanner.h"
-#include "csv-contextual-data-record-scanner.h"
+#include "scanner/csv-scanner/csv-scanner.h"
+#include "string-list.h"
 #include "messages.h"
+#include "cfg.h"
+
 #include <string.h>
 
-void
-contextual_data_record_scanner_free(ContextualDataRecordScanner *self)
+struct _ContextualDataRecordScanner
 {
-  if (self && self->free_fn)
-    self->free_fn(self);
+  ContextualDataRecord last_record;
+  GlobalConfig *cfg;
+  CSVScanner scanner;
+  CSVScannerOptions options;
+  gchar *filename;
+  const gchar *name_prefix;
+};
+
+
+static gchar *
+_csv_scanner_dup_current_value_with_prefix(CSVScanner *line_scanner,
+                                           const gchar *prefix)
+{
+  return g_strdup_printf("%s%s", prefix ? prefix : "",
+                         csv_scanner_get_current_value(line_scanner));
 }
 
-void
-contextual_data_record_scanner_set_name_prefix(ContextualDataRecordScanner *
-                                               self, const gchar *prefix)
+static gboolean
+_fetch_next_with_prefix(ContextualDataRecordScanner *record_scanner,
+                        GString **target, const gchar *prefix)
 {
-  self->name_prefix = prefix;
-}
-
-ContextualDataRecordScanner *
-create_contextual_data_record_scanner_by_type(GlobalConfig *cfg, const gchar *filename, const gchar *type)
-{
-  ContextualDataRecordScanner *scanner = NULL;
-
-  if (type == NULL)
-    return NULL;
-
-  if (!strcmp(type, "csv"))
+  if (!csv_scanner_scan_next(&record_scanner->scanner))
     {
-      scanner = csv_contextual_data_record_scanner_new(cfg, filename);
+      msg_error("add-contextual-data(): error parsing CSV file, expecting an additional column which was not found. Expecting (selector, name, value) triplets",
+                evt_tag_str("filename", record_scanner->filename),
+                evt_tag_str("target", csv_scanner_get_current_name(&record_scanner->scanner)));
+      return FALSE;
     }
+  gchar *next = _csv_scanner_dup_current_value_with_prefix(&record_scanner->scanner, prefix);
+  *target = g_string_new(next);
+  g_free(next);
 
-  if (!scanner)
-    msg_warning("Unknown ContextualDataRecordScanner",
-                evt_tag_str("type", type));
+  return TRUE;
+}
 
+static gboolean
+_fetch_next_without_prefix(ContextualDataRecordScanner *
+                           record_scanner, GString **target)
+{
+  return _fetch_next_with_prefix(record_scanner, target, NULL);
+}
 
-  return scanner;
+static gboolean
+_is_whole_record_parsed(ContextualDataRecordScanner *csv_record_scanner)
+{
+  if (!csv_scanner_scan_next(&csv_record_scanner->scanner) &&
+      csv_scanner_is_scan_complete(&csv_record_scanner->scanner))
+    return TRUE;
+
+  msg_error("add-contextual-data(): extra data found at the end of line, expecting (selector, name, value) triplets",
+            evt_tag_str("filename", csv_record_scanner->filename));
+  return FALSE;
+}
+
+static gboolean
+_get_next_record(ContextualDataRecordScanner *self, const gchar *input, ContextualDataRecord *record)
+{
+  GString *value_template;
+  GString *name;
+  GError *error = NULL;
+
+  csv_scanner_init(&self->scanner, &self->options, input);
+
+  if (!_fetch_next_without_prefix(self, &record->selector))
+    goto error;
+
+  if (!_fetch_next_with_prefix(self, &name, self->name_prefix))
+    goto error;
+
+  record->value_handle = log_msg_get_value_handle(name->str);
+  g_string_free(name, TRUE);
+
+  if (!_fetch_next_without_prefix(self, &value_template))
+    goto error;
+
+  record->value = log_template_new(self->cfg, NULL);
+
+  if (cfg_is_config_version_older(self->cfg, VERSION_VALUE_3_21) &&
+      strchr(value_template->str, '$') != NULL)
+    {
+      msg_warning("WARNING: the value field in add-contextual-data() CSV files has been changed "
+                  "to be a template starting with " VERSION_3_21 ". You are using an older config "
+                  "version and your CSV file contains a '$' character in this field, which needs "
+                  "to be escaped as '$$' once you change your @version declaration in the "
+                  "configuration. This message means that this string is now assumed to be a "
+                  "literal (non-template) string for compatibility",
+                  evt_tag_str("selector", record->selector->str),
+                  evt_tag_str("name", log_msg_get_value_name(record->value_handle, NULL)),
+                  evt_tag_str("value", value_template->str));
+      log_template_compile_literal_string(record->value, value_template->str);
+    }
+  else
+    {
+      if (!log_template_compile(record->value, value_template->str, &error))
+        {
+          msg_error("add-contextual-data(): error compiling template",
+                   evt_tag_str("selector", record->selector->str),
+                   evt_tag_str("name", log_msg_get_value_name(record->value_handle, NULL)),
+                   evt_tag_str("value", value_template->str));
+          g_string_free(value_template, TRUE);
+          goto error;
+        }
+    }
+  g_string_free(value_template, TRUE);
+
+  if (!_is_whole_record_parsed(self))
+    goto error;
+
+  return TRUE;
+error:
+  msg_error("add-contextual-data(): the failing line is",
+            evt_tag_str("filename", self->filename),
+            evt_tag_str("input", input));
+  return FALSE;
 }
 
 ContextualDataRecord *
 contextual_data_record_scanner_get_next(ContextualDataRecordScanner *self, const gchar *input)
 {
-  if (!self->get_next)
-    return NULL;
-
   contextual_data_record_init(&self->last_record);
-  if (!self->get_next(self, input, &self->last_record))
+  if (!_get_next_record(self, input, &self->last_record))
     {
       contextual_data_record_clean(&self->last_record);
       return NULL;
@@ -77,7 +160,37 @@ contextual_data_record_scanner_get_next(ContextualDataRecordScanner *self, const
 }
 
 void
-contextual_data_record_scanner_init(ContextualDataRecordScanner *self, GlobalConfig *cfg)
+contextual_data_record_scanner_set_name_prefix(ContextualDataRecordScanner *
+                                               self, const gchar *prefix)
 {
+  self->name_prefix = prefix;
+}
+
+void
+contextual_data_record_scanner_free(ContextualDataRecordScanner *self)
+{
+  csv_scanner_options_clean(&self->options);
+  csv_scanner_deinit(&self->scanner);
+  g_free(self->filename);
+  g_free(self);
+}
+
+ContextualDataRecordScanner *
+contextual_data_record_scanner_new(GlobalConfig *cfg, const gchar *filename)
+{
+  ContextualDataRecordScanner *self =
+    g_new0(ContextualDataRecordScanner, 1);
+
   self->cfg = cfg;
+
+  csv_scanner_options_set_delimiters(&self->options, ",");
+  csv_scanner_options_set_quote_pairs(&self->options, "\"\"''");
+  const gchar *column_array[] = { "selector", "name", "value", NULL };
+  csv_scanner_options_set_columns(&self->options,
+                                  string_array_to_list(column_array));
+  csv_scanner_options_set_flags(&self->options, CSV_SCANNER_STRIP_WHITESPACE);
+  csv_scanner_options_set_dialect(&self->options, CSV_SCANNER_ESCAPE_DOUBLE_CHAR);
+  self->filename = g_strdup(filename);
+
+  return self;
 }
