+++ modules/add-contextual-data/tests/test_context_info_db.c
@@ -219,7 +219,7 @@ _assert_import_csv_with_single_selector(gchar *csv_content, gchar *selector_to_c
   FILE *fp = fmemopen(csv_content, strlen(csv_content), "r");
   ContextInfoDB *db = context_info_db_new(FALSE);
   ContextualDataRecordScanner *scanner =
-    create_contextual_data_record_scanner_by_type(configuration, "dummy.csv", "csv");
+    contextual_data_record_scanner_new(configuration, "dummy.csv");
 
   cr_assert(context_info_db_import(db, fp, scanner),
             "Failed to import valid CSV file.");
@@ -262,7 +262,7 @@ Test(add_contextual_data, test_import_with_valid_csv)
   FILE *fp = fmemopen(csv_content, sizeof(csv_content), "r");
   ContextInfoDB *db = context_info_db_new(FALSE);
   ContextualDataRecordScanner *scanner =
-    create_contextual_data_record_scanner_by_type(configuration, "dummy.csv", "csv");
+    contextual_data_record_scanner_new(configuration, "dummy.csv");
 
   cr_assert(context_info_db_import(db, fp, scanner),
             "Failed to import valid CSV file.");
@@ -342,7 +342,7 @@ Test(add_contextual_data, test_import_with_invalid_csv_content)
   ContextInfoDB *db = context_info_db_new(FALSE);
 
   ContextualDataRecordScanner *scanner =
-    create_contextual_data_record_scanner_by_type(configuration, "dummy.csv", "csv");
+    contextual_data_record_scanner_new(configuration, "dummy.csv");
 
   cr_assert_not(context_info_db_import(db, fp, scanner),
                 "Successfully import an invalid CSV file.");
@@ -364,7 +364,7 @@ Test(add_contextual_data, test_import_with_csv_contains_invalid_line)
   ContextInfoDB *db = context_info_db_new(FALSE);
 
   ContextualDataRecordScanner *scanner =
-    create_contextual_data_record_scanner_by_type(configuration, "dummy.csv", "csv");
+    contextual_data_record_scanner_new(configuration, "dummy.csv");
 
   cr_assert_not(context_info_db_import(db, fp, scanner),
                 "Successfully import an invalid CSV file.");
@@ -429,7 +429,7 @@ ParameterizedTest(struct TestNVPairPrefix *param, add_contextual_data, test_impo
   FILE *fp = fmemopen(csv_content, sizeof(csv_content), "r");
   ContextInfoDB *db = context_info_db_new(FALSE);
   ContextualDataRecordScanner *scanner =
-    create_contextual_data_record_scanner_by_type(configuration, "dummy.csv", "csv");
+    contextual_data_record_scanner_new(configuration, "dummy.csv");
 
   contextual_data_record_scanner_set_name_prefix(scanner, param->prefix);
 
@@ -454,7 +454,7 @@ Test(add_contextual_data, test_ignore_case_on)
   ContextInfoDB *db = context_info_db_new(TRUE);
 
   ContextualDataRecordScanner *scanner =
-    create_contextual_data_record_scanner_by_type(configuration, "dummy.csv", "csv");
+    contextual_data_record_scanner_new(configuration, "dummy.csv");
 
   cr_assert(context_info_db_import(db, fp, scanner),
             "Failed to import valid CSV file.");
@@ -477,7 +477,7 @@ Test(add_contextual_data, test_ignore_case_off)
   ContextInfoDB *db = context_info_db_new(FALSE);
 
   ContextualDataRecordScanner *scanner =
-    create_contextual_data_record_scanner_by_type(configuration, "dummy.csv", "csv");
+    contextual_data_record_scanner_new(configuration, "dummy.csv");
 
   cr_assert(context_info_db_import(db, fp, scanner),
             "Failed to import valid CSV file.");
@@ -504,7 +504,7 @@ Test(add_contextual_data, test_selected_nvpairs_when_ignore_case_on)
 
   ContextInfoDB *db = context_info_db_new(TRUE);
   ContextualDataRecordScanner *scanner =
-    create_contextual_data_record_scanner_by_type(configuration, "dummy.csv", "csv");
+    contextual_data_record_scanner_new(configuration, "dummy.csv");
 
   cr_assert(context_info_db_import(db, fp, scanner),
             "Failed to import valid CSV file.");
